backend/requirements.txt
# Основные зависимости
setuptools>=69.0.3
wheel>=0.42.0
fastapi>=0.109.2
uvicorn>=0.27.1
python-multipart>=0.0.7
pydantic>=2.6.1
pydantic-settings>=2.1.0

# ML и обработка документов
transformers>=4.37.2
torch>=2.2.0
pdfminer.six>=20231228
python-docx>=1.1.0

# Утилиты
python-magic>=0.4.27
aiofiles>=23.2.1
numpy>=1.26.3
--find-links https://download.pytorch.org/whl/torch_stable.html
httpx 